requirements.txt
streamlit
pandas
numpy
pillow
giotto-tda
matplotlib